requirements.txt
flask
joblib
numpy
cassandra-driver
python-whois
bs4
requests
dnspython
scikit-learn